% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comparison_of_interest}
\alias{comparison_of_interest}
\title{Cell-cell interation of interest}
\format{
A character variable
}
\usage{
comparison_of_interest
}
\description{
Cell-cell interation of interest
}
\keyword{datasets}
