% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comparison.R
\name{make_comparison}
\alias{make_comparison}
\alias{run_survival}
\title{Survival analysis for scACCorDiON results}
\usage{
make_comparison(
  comparison_oi,
  lr_data,
  tcga_exp_data,
  tcga_clinical_data,
  selection_method = c("limma", "prauc"),
  custom_selection = NULL,
  n_lr_selected = 10,
  which_get = c("ligand", "receptor"),
  is_signif = TRUE,
  clinical_vars_in_model = NULL
)

run_survival(
  comparison_oi,
  lr_data,
  tcga_exp_data,
  tcga_clinical_data,
  selection_method = c("limma", "prauc"),
  custom_selection = NULL,
  n_lr_selected = 10,
  which_get = c("ligand", "receptor"),
  is_signif = TRUE,
  clinical_vars_in_model = NULL
)
}
\arguments{
\item{comparison_oi}{TODO}

\item{lr_data}{TODO}

\item{tcga_exp_data}{TODO}

\item{tcga_clinical_data}{TODO}

\item{selection_method}{TODO}

\item{custom_selection}{TODO}

\item{n_lr_selected}{TODO}

\item{which_get}{TODO}

\item{is_signif}{TODO}

\item{clinical_vars_in_model}{TODO}
}
\value{
TODO
}
\description{
Survival analysis for scACCorDiON results
}
