% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paad_tcga_clinical_data}
\alias{paad_tcga_clinical_data}
\title{PAAD TCGA clinical data}
\format{
A tibble
}
\usage{
paad_tcga_clinical_data
}
\description{
PAAD TCGA clinical data
}
\keyword{datasets}
