% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paad_tcga_expression_data}
\alias{paad_tcga_expression_data}
\title{PAAD TCGA expression data}
\format{
A tibble
}
\usage{
paad_tcga_expression_data
}
\description{
PAAD TCGA expression data
}
\keyword{datasets}
