% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdac_cci_data}
\alias{pdac_cci_data}
\title{PDAC cell-cell interaction data from scACCorDiON}
\format{
A tibble
}
\usage{
pdac_cci_data
}
\description{
PDAC cell-cell interaction data from scACCorDiON
}
\keyword{datasets}
