% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scACCorDiON.su-package.R
\docType{package}
\name{scACCorDiON.su-package}
\alias{scACCorDiON.su}
\alias{scACCorDiON.su-package}
\title{scACCorDiON.su: Survival Analysis on scACCorDiON results}
\description{
Produce survival analysis on ligand receptor results from scACCorDiON.
}
\author{
\strong{Maintainer}: Tiago Maié \email{tiagomaie+Rdev@hotmail.com}

Authors:
\itemize{
  \item James Nagai \email{james.nagai+Rdev@gmail.com}
}

}
\keyword{internal}
